fix: 도메인객체와 DTO 분리
