+++ README.md
@@ -94,7 +94,7 @@
 
 ### 리뷰반영사항 - 1
 - [x] Car::에러메시지 상수적용
-- [ ] RaceBoard::도메인객체와 DTO 분리
+- [x] RaceBoard::도메인객체와 DTO 분리
 - [ ] ResultView::우승자 출력시 우승자를 파라미터로 추출
 - [ ] CarTest::네이밍 공통화
 - [ ] ResultView::불필요한 메서드추출 병합
