+++ src/main/kotlin/racingcar/domain/dto/CarPosition.kt
@@ -1,0 +1,14 @@
+package racingcar.domain.dto
+
+import racingcar.domain.Car
+
+data class CarPosition(
+    val name: String,
+    val position: Int,
+) {
+    companion object {
+        fun from(car: Car): CarPosition {
+            return CarPosition(car.name, car.position)
+        }
+    }
+}
