+++ src/main/kotlin/racingcar/domain/dto/Winner.kt
@@ -1,0 +1,5 @@
+package racingcar.domain.dto
+
+data class Winner(
+    val name: String,
+)
